refactor error functions to separate file
